+++ csrc/utils.h
@@ -514,210 +514,4 @@ class KernelIndexTypeCompute {
   int64_t tensor_most_positive_index_ = 0;
 };
 
-namespace has_operator_impl {
-
-struct HasOperatorHelper {};
-
-struct IsNullaryFunc {
-  template <typename T>
-  static constexpr auto check(int) -> decltype((std::declval<T>()()), true) {
-    return true;
-  }
-
-  template <typename T>
-  static constexpr bool check(long) {
-    return false;
-  }
-};
-
-struct HasArrowOperator {
-  template <typename T>
-  static constexpr auto check(int)
-      -> decltype((std::declval<decltype(&T::operator->)>()), true) {
-    return true;
-  }
-
-  template <typename T>
-  static constexpr bool check(long) {
-    return false;
-  }
-};
-
-struct TrueType {
-  static constexpr bool value() {
-    return true;
-  }
-};
-
-struct FalseType {
-  static constexpr bool value() {
-    return false;
-  }
-};
-
-template <typename T>
-struct HasOperator {
-  constexpr operator HasOperatorHelper() const {
-    return {};
-  }
-
-  template <typename T1>
-  constexpr auto operator=(HasOperator<T1>) const
-      -> decltype((std::declval<T>() = std::declval<T1>()), true) {
-    return true;
-  }
-
-  constexpr bool operator=(HasOperatorHelper) const {
-    return false;
-  }
-
-  template <typename... Ts>
-  constexpr auto operator()(HasOperator<Ts>... args) const
-      -> decltype((std::declval<T>()(std::declval<Ts>()...)), true) {
-    return true;
-  }
-
-  constexpr bool operator()(HasOperatorHelper) const {
-    return false;
-  }
-
-  template <typename... Ts>
-  constexpr bool operator()(HasOperatorHelper, Ts... args) const {
-    return false && operator()(args...);
-  }
-
-  template <
-      typename T1 = int,
-      std::enable_if_t<!IsNullaryFunc::check<T>(int{}), T1> = 0>
-  constexpr bool operator()() const {
-    return false;
-  }
-
-  template <typename T1>
-  constexpr auto operator[](HasOperator<T1>) const
-      -> decltype((std::declval<T>()[std::declval<T1>()]), true) {
-    return true;
-  }
-
-  constexpr bool operator[](HasOperatorHelper) const {
-    return false;
-  }
-
-  template <
-      typename T1 = int,
-      std::enable_if_t<HasArrowOperator::check<T>(int{}), T1> = 0>
-  constexpr auto operator->() const -> TrueType* {
-    return nullptr;
-  }
-
-  template <
-      typename T1 = int,
-      std::enable_if_t<!HasArrowOperator::check<T>(int{}), T1> = 0>
-  constexpr auto operator->() const -> FalseType* {
-    return nullptr;
-  }
-};
-
-#define DEFINE_UNARY_OP(op)                       \
-  template <typename T1>                          \
-  constexpr auto operator op(HasOperator<T1>)     \
-      ->decltype(op std::declval<T1>(), true) {   \
-    return true;                                  \
-  }                                               \
-                                                  \
-  constexpr bool operator op(HasOperatorHelper) { \
-    return false;                                 \
-  }
-
-#define DEFINE_UNARY_SUFFIX_OP(op)                     \
-  template <typename T1>                               \
-  constexpr auto operator op(HasOperator<T1>, int)     \
-      ->decltype(std::declval<T1>() op, true) {        \
-    return true;                                       \
-  }                                                    \
-                                                       \
-  constexpr bool operator op(HasOperatorHelper, int) { \
-    return false;                                      \
-  }
-
-#define DEFINE_BINARY_OP(op)                                         \
-  template <typename T1, typename T2>                                \
-  constexpr auto operator op(HasOperator<T1>, HasOperator<T2>)       \
-      ->decltype((std::declval<T1>() op std::declval<T2>()), true) { \
-    return true;                                                     \
-  }                                                                  \
-                                                                     \
-  constexpr bool operator op(HasOperatorHelper, HasOperatorHelper) { \
-    return false;                                                    \
-  }
-
-// Unary operators
-DEFINE_UNARY_OP(+);
-DEFINE_UNARY_OP(-);
-DEFINE_UNARY_OP(~);
-DEFINE_UNARY_OP(!);
-DEFINE_UNARY_OP(++);
-DEFINE_UNARY_OP(--);
-DEFINE_UNARY_SUFFIX_OP(++);
-DEFINE_UNARY_SUFFIX_OP(--);
-DEFINE_UNARY_OP(*);
-DEFINE_UNARY_OP(&);
-
-// Binary operators
-DEFINE_BINARY_OP(+);
-DEFINE_BINARY_OP(-);
-DEFINE_BINARY_OP(*);
-DEFINE_BINARY_OP(/);
-DEFINE_BINARY_OP(%);
-DEFINE_BINARY_OP(&);
-DEFINE_BINARY_OP(|);
-DEFINE_BINARY_OP(^);
-DEFINE_BINARY_OP(&&);
-DEFINE_BINARY_OP(||);
-DEFINE_BINARY_OP(<<);
-DEFINE_BINARY_OP(>>);
-DEFINE_BINARY_OP(==);
-DEFINE_BINARY_OP(!=);
-DEFINE_BINARY_OP(<);
-DEFINE_BINARY_OP(>);
-DEFINE_BINARY_OP(<=);
-DEFINE_BINARY_OP(>=);
-
-// Assignment operators
-DEFINE_BINARY_OP(+=);
-DEFINE_BINARY_OP(-=);
-DEFINE_BINARY_OP(*=);
-DEFINE_BINARY_OP(/=);
-DEFINE_BINARY_OP(%=);
-DEFINE_BINARY_OP(&=);
-DEFINE_BINARY_OP(|=);
-DEFINE_BINARY_OP(^=);
-DEFINE_BINARY_OP(<<=);
-DEFINE_BINARY_OP(>>=);
-
-#undef DEFINE_UNARY_OP
-#undef DEFINE_UNARY_SUFFIX_OP
-#undef DEFINE_BINARY_OP
-
-// comma operator
-template <typename T1, typename T2>
-constexpr auto operator,(HasOperator<T1>, HasOperator<T2>)
-    -> decltype((std::declval<T1>(), std::declval<T2>()), true) {
-  return true;
-}
-
-constexpr bool operator,(HasOperatorHelper, HasOperatorHelper) {
-  return false;
-}
-
-// TODO: overload the following operators:
-// <=> (requires C++20)
-// ->* (requires pointer, unary and must be member)
-// -> (requires pointer, unary and must be member)
-
-} // namespace has_operator_impl
-
-template <typename T>
-constexpr has_operator_impl::HasOperator<T> has_operator;
-
 } // namespace nvfuser
